requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
bs4==0.0.2
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cssselect==1.2.0; python_version >= '3.7'
et-xmlfile==1.1.0; python_version >= '3.6'
google-api-core[grpc]==2.19.0; python_version >= '3.7'
google-auth==2.30.0; python_version >= '3.7'
google-cloud-vision==3.4.5; python_version >= '3.7'
googleapis-common-protos==1.63.1; python_version >= '3.7'
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.7; python_version >= '3.5'
lxml==4.9.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdownify==0.12.1
openpyxl==3.1.4; python_version >= '3.8'
pillow==10.0.1; python_version >= '3.8'
proto-plus==1.23.0; python_version >= '3.6'
protobuf==4.25.3; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pypandoc==1.13; python_version >= '3.6'
pypandoc-binary==1.13; python_version >= '3.6'
python-dotenv==1.0.1; python_version >= '3.8'
readability-lxml==0.8.1
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
xlrd==2.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytest==8.2.2; python_version >= '3.8'